Move ValidatesLicenseTypeMixin to common module
